# Remove no_magic support; which was not working for dbapi