--- a/gramps/plugins/lib/libgedcom.py
+++ b/gramps/plugins/lib/libgedcom.py
@@ -2737,9 +2737,8 @@ class GedcomParser(UpdateCallback):
           0 TRLR                                          {1:1}
 
         """
-        no_magic = self.maxpeople < 1000
-        with DbTxn(_("GEDCOM import"), self.dbase, not use_trans,
-                   no_magic=no_magic) as self.trans:
+        with DbTxn(_("GEDCOM import"), self.dbase,
+                   not use_trans) as self.trans:
 
             self.dbase.disable_signals()
             self.__parse_header_head()